requirements.txt
python-telegram-bot>=20.0
PyMuPDF>=1.24.0
